Feature/material count type

物料数量类型有Long改为Double

See merge request !123

+++ doctor-warehouse-api/src/main/java/io/terminus/doctor/warehouse/dto/DoctorMaterialConsumeProviderDto.java
@@ -53,7 +53,7 @@ public class DoctorMaterialConsumeProviderDto implements Serializable{
 
     private String staffName;  //饲养员信息
 
-    private Long count;
+    private Double count;
 
     private Long unitPrice; // 单价, 单位为"分"
