+++ doctor-warehouse/src/main/java/io/terminus/doctor/warehouse/handler/consume/DoctorMaterialAvgConsumerHandler.java
@@ -41,14 +41,14 @@ public class DoctorMaterialAvgConsumerHandler implements IHandler{
 
     @Override
     public void handle(DoctorMaterialConsumeProviderDto dto, Map<String, Object> context) throws RuntimeException {
-        Long lotNumber = (Long) context.get("lotNumber");
+        Double lotNumber = (Double) context.get("lotNumber");
         DoctorMaterialConsumeAvg doctorMaterialConsumeAvg = doctorMaterialConsumeAvgDao.queryByIds(dto.getFarmId(), dto.getWareHouseId(), dto.getMaterialTypeId());
         if(isNull(doctorMaterialConsumeAvg)){
             // create consume avg
             doctorMaterialConsumeAvg = DoctorMaterialConsumeAvg.builder()
                     .farmId(dto.getFarmId()).wareHouseId(dto.getWareHouseId()).materialId(dto.getMaterialTypeId())
                     .consumeDate(DateTime.now().withTimeAtStartOfDay().toDate()).consumeCount(dto.getCount())
-                    .consumeAvgCount(0l)
+                    .consumeAvgCount(0D)
                     .build();
 
             if(Objects.equals(dto.getType(), WareHouseType.FEED.getKey())){
