+++ doctor-web/src/main/java/io/terminus/doctor/web/front/warehouse/dto/DoctorConsumeProviderInputDto.java
@@ -27,7 +27,7 @@ public class DoctorConsumeProviderInputDto implements Serializable{
 
     private Long materialId;    // 对应的物料Id
 
-    private Long count;  //对应的数量
+    private Double count;  //对应的数量
 
     private Long unitPrice; // 单价, 单位为"分"
